requirements.txt
pandas==2.3.3
numpy==2.2.6
simpleitk==2.5.2
nibabel==5.3.2
scikit-image==0.25.0
TotalSegmentator==2.11.0
openpyxl==3.1.5
uvicorn==0.37.0
fastapi==0.118.0
pydantic==2.11.9
python-jose==3.5.0
bcrypt==4.0.1 
passlib[bcrypt]
aiofiles==24.1.0